rename lib folder
