+++ website/frontend/src/components/blocks/UploadMetaForm/helpers/getMetaValues.ts
